# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/griddlehq/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/indexes": {
      "get": {
        "tags": [
          "indices"
        ],
        "summary": "List all indices",
        "operationId": "list_indices",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (null = return all unpaginated)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "hitsPerPage",
            "in": "query",
            "description": "Number of indices per page (default 100, max 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all indices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIndicesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "indices"
        ],
        "summary": "Create a new index",
        "operationId": "create_index",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Index created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Index already exists"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIndexRequest": {
        "type": "object",
        "required": [
          "uid"
        ],
        "properties": {
          "schema": {
            "$ref": "#/components/schemas/IndexSchema"
          },
          "uid": {
            "type": "string"
          }
        }
      },
      "CreateIndexResponse": {
        "type": "object",
        "required": [
          "uid",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          }
        }
      },
...
      "IndexSchema": {
        "type": "object",
        "properties": {
          "facet_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "float_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "integer_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "text_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ListIndexItem": {
        "type": "object",
        "description": "Single index entry returned by the list-indices endpoint.\n\nField names are serialized in camelCase to match the Algolia REST API wire format.\n`replicas`, `primary`, and `virtual_replica` are omitted from the response when `None`.",
        "required": [
          "name",
          "createdAt",
          "updatedAt",
          "entries",
          "dataSize",
          "fileSize",
          "lastBuildTimeS",
          "numberOfPendingTasks",
          "pendingTask"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "dataSize": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "entries": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "lastBuildTimeS": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "numberOfPendingTasks": {
            "type": "integer",
            "minimum": 0
          },
          "pendingTask": {
            "type": "boolean"
          },
          "primary": {
            "type": [
              "string",
              "null"
            ]
          },
          "replicas": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string"
          },
          "virtual": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
      "ListIndicesResponse": {
        "type": "object",
        "required": [
          "items",
          "nbPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListIndexItem"
            }
          },
          "nbPages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}